+++ packages/core/crates/core-network/src/ping.rs
@@ -256,7 +256,6 @@ pub mod wasm {
     use js_sys::JsString;
     use std::str::FromStr;
     use wasm_bindgen::prelude::*;
-    use wasm_bindgen::JsCast;
 
     #[wasm_bindgen]
     struct WasmPingApi {
@@ -346,25 +345,34 @@ pub mod wasm {
                     Box::pin(async move {
                         let this = JsValue::null();
                         let data: JsValue = js_sys::Uint8Array::from(msg.as_ref()).into();
-                        let peer: JsValue = JsString::from(peer.as_str()).into();
+                        let peer: JsValue = <JsValue as From<String>>::from(peer);
 
                         // call a send_msg_cb producing a JS promise that is further converted to a Future
                         // holding the reply of the pinged peer for the ping message.
                         match self.send_msg_cb.call2(&this, &data, &peer) {
                             Ok(r) => {
                                 let promise = js_sys::Promise::from(r);
-                                let data = wasm_bindgen_futures::JsFuture::from(promise)
-                                    .await
-                                    .map(|x| js_sys::Array::from(x.as_ref()).get(0))
-                                    .map(|x| js_sys::Uint8Array::new(&x).to_vec().into_boxed_slice())
-                                    .map_err(|x| {
-                                        x.dyn_ref::<JsString>()
-                                            .map_or("Failed to send ping message".to_owned(), |x| -> String {
-                                                x.into()
-                                            })
-                                    });
-
-                                data
+                                match wasm_bindgen_futures::JsFuture::from(promise).await {
+                                    Ok(x) => {
+                                        if x.is_undefined() {
+                                            Err("Failed to send ping message".into())
+                                        } else {
+                                            debug!("transport returned {:?}", x);
+                                            let arr = js_sys::Array::from(x.as_ref());
+                                            if arr.length() > 0 {
+                                                Ok(js_sys::Uint8Array::from(arr.get(0)).to_vec().into_boxed_slice())
+                                            } else {
+                                                error!("transport has returned an empty response");
+                                                Err("Empty response returned from ping transport".into())
+                                            }
+                                        }
+                                    }
+                                    Err(e) => {
+                                        error!("Failed to send ping message");
+                                        error!("{:?}", e);
+                                        Err("Failed to send ping message".into())
+                                    }
+                                }
                             }
                             Err(e) => {
                                 error!(